refactor(rolldown_sourcemap): tweak methods, namings and reduce some unnecessary memory allocations
