+++ crates/rolldown/src/ecmascript/ecma_generator.rs
@@ -122,13 +122,13 @@ impl Generator for EcmaGenerator {
         .await?
     };
 
-    let concat_source = match ctx.options.format {
+    let source_joiner = match ctx.options.format {
       OutputFormat::Esm => {
         render_esm(ctx, rendered_module_sources, banner, footer, intro, outro, hashbang)
       }
       OutputFormat::Cjs => {
         match render_cjs(ctx, rendered_module_sources, banner, footer, intro, outro, hashbang) {
-          Ok(concat_source) => concat_source,
+          Ok(source_joiner) => source_joiner,
           Err(errors) => return Ok(Err(errors)),
         }
       }
@@ -137,19 +137,19 @@ impl Generator for EcmaGenerator {
       }
       OutputFormat::Iife => {
         match render_iife(ctx, rendered_module_sources, banner, footer, intro, outro, hashbang) {
-          Ok(concat_source) => concat_source,
+          Ok(source_joiner) => source_joiner,
           Err(errors) => return Ok(Err(errors)),
         }
       }
       OutputFormat::Umd => {
         match render_umd(ctx, rendered_module_sources, banner, footer, intro, outro) {
-          Ok(concat_source) => concat_source,
+          Ok(source_joiner) => source_joiner,
           Err(errors) => return Ok(Err(errors)),
         }
       }
     };
 
-    let (content, mut map) = concat_source.content_and_sourcemap();
+    let (content, mut map) = source_joiner.join();
 
     // Here file path is generated by chunk file name template, it maybe including path segments.
     // So here need to read it's parent directory as file_dir.
