+++ crates/rolldown/src/utils/render_ecma_module.rs
@@ -1,5 +1,5 @@
 use rolldown_common::{NormalModule, NormalizedBundlerOptions};
-use rolldown_sourcemap::{collapse_sourcemaps, lines_count, RawSource, Source, SourceMapSource};
+use rolldown_sourcemap::{collapse_sourcemaps, lines_count, Source, SourceMapSource};
 
 use crate::types::generator::CodegenArtifact;
 
@@ -12,10 +12,8 @@ pub fn render_ecma_module(
     None
   } else {
     let mut sources: Vec<Box<dyn rolldown_sourcemap::Source + Send>> = vec![];
-    sources.push(Box::new(RawSource::new(format!(
-      "//#region {debug_module_id}",
-      debug_module_id = module.debug_id
-    ))));
+    sources
+      .push(Box::new(format!("//#region {debug_module_id}", debug_module_id = module.debug_id)));
 
     let enable_sourcemap = options.sourcemap.is_some() && !module.is_virtual();
 
@@ -38,13 +36,13 @@ pub fn render_ecma_module(
         let lines_count = lines_count(&render_output.code);
         sources.push(Box::new(SourceMapSource::new(render_output.code, sourcemap, lines_count)));
       } else {
-        sources.push(Box::new(RawSource::new(render_output.code)));
+        sources.push(Box::new(render_output.code));
       }
     } else {
-      sources.push(Box::new(RawSource::new(render_output.code)));
+      sources.push(Box::new(render_output.code));
     }
 
-    sources.push(Box::new(RawSource::new("//#endregion".to_string())));
+    sources.push(Box::new("//#endregion"));
 
     Some(sources)
   }
